+++ CHANGELOG.md
@@ -5,6 +5,7 @@
  - Added effects to fill and blink.
  - Fix boon spell's icon.
  - Add effects to disintegrate, shrink, fire, frost, disarm.
+ - Put a hack in to work-around annoying double-interact event breaking fill.
  
 ## 2.7.1